change file name to jsx
